Remove mqtt server, add fauxmo service
mqtt server no longer needed, the new fauxmo service can make a POST
request to the openhab REST API to change item state.

Patches added to disable homeassistant code paths because I won't be
using homeassistant and the homeassistant dependency was too hard to
add.

Fauxmo requires the UDP port 1900 to be open for UPnP broadcast
discovery from the Amazon Echo. And then a port needs to be added for
each Fauxmo device. Fauxmo pretends to be a Wemo smart switch and allows
On/Off commands.

This approach is a lot more responsive than using the built-in Echo
Philips Hue commands.